Update 74HC595_register_and_MAX6675.h
change the data type (float => double)

+++ src/74HC595_register_and_MAX6675.h
@@ -22,5 +22,5 @@ namespace 74HC595
 namespace MAX6675
 {
   void setting_pin_values_for_SPI();
-  float readCelsius();
+  double readCelsius();
 }
